feat: create utility to generate urls, separate function to output logins
