+++ webpack.config.js
@@ -1,6 +1,6 @@
 const path = require('path');
 module.exports = {
-  entry: './src/index.ts',
+  entry: './src/index.js',
   devtool: 'inline-source-map',
   module: {
     rules: [
